backend/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anthropic==0.59.0
anyio==4.9.0
attrs==25.3.0
backports.asyncio.runner==1.2.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.97.1
openapi==2.0.0
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
packaging==25.0
pathable==0.4.4
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
swagger-spec-validator==3.0.4
swagger_ui_bundle==1.1.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0